+++ gito/gh_api.py
@@ -2,10 +2,7 @@ import os
 import logging
 
 import requests
-from ghapi.all import GhApi
 from fastcore.basics import AttrDict  # objects returned by ghapi
-
-from .constants import HTML_CR_COMMENT_MARKER
 
 
 def resolve_gh_token(token_or_none):
@@ -42,39 +39,6 @@ def post_gh_comment(
     else:
         logging.error(f"Failed to post comment: {resp.status_code} {resp.reason}\n{resp.text}")
         return False
-
-
-def collapse_gh_outdated_cr_comments(
-    gh_repository: str,
-    pr_or_issue_number: int,
-    token: str = None
-):
-    """
-    Collapse outdated code review comments in a GitHub pull request or issue.
-    """
-    logging.info(f"Collapsing outdated comments in {gh_repository} #{pr_or_issue_number}...")
-
-    token = resolve_gh_token(token)
-    owner, repo = gh_repository.split('/')
-    api = GhApi(owner, repo, token=token)
-
-    comments = api.issues.list_comments(pr_or_issue_number)
-    review_marker = HTML_CR_COMMENT_MARKER
-    collapsed_title = "🗑️ <s>Outdated Code Review by Gito</s>"
-    collapsed_marker = f"<summary>{collapsed_title}</summary>"
-    outdated_comments = [
-        c for c in comments
-        if c.body and review_marker in c.body and collapsed_marker not in c.body
-    ][:-1]
-    if not outdated_comments:
-        logging.info("No outdated comments found")
-        return
-    for comment in outdated_comments:
-        logging.info(f"Collapsing comment {comment.id}...")
-        new_body = f"<details>\n<summary>{collapsed_title}</summary>\n\n{comment.body}\n</details>"
-        api.issues.update_comment(comment.id, new_body)
-        hide_gh_comment(comment.node_id, token)
-    logging.info("All outdated comments collapsed successfully.")
 
 
 def hide_gh_comment(
